Move spec-specific parsing back into parser.py (per feedback)
